:recycle: Rename generate-confidentiality-agreement serializer folder to generate-confidentiality-receipt.
