--- api/src/serializers/information-sharing-agreements/generate-confidentiality-agreement/index.ts
@@ -1,4 +1,0 @@
-export {
-  CreateSerializer,
-  type InformationSharingAgreementAsConfidentialityAgreement as AsConfidentialityReceipt,
-} from "./create-serializer"
