+++ api/src/serializers/information-sharing-agreements/generate-confidentiality-receipt/index.ts
@@ -1,0 +1,4 @@
+export {
+  CreateSerializer,
+  type InformationSharingAgreementAsConfidentialityReceipt as AsConfidentialityReceipt,
+} from "./create-serializer"
